reorg the beam sql examples in their own project under examples
