+++ settings.gradle.kts
@@ -78,6 +78,7 @@ include(":examples:java:cdap:hubspot")
 include(":examples:java:cdap:salesforce")
 include(":examples:java:cdap:servicenow")
 include(":examples:java:cdap:zendesk")
+include(":examples:java:sql")
 include(":examples:java:webapis")
 include(":examples:kotlin")
 include(":examples:multi-language")
